Extract reflection utils into a single `Invokables` class.
